Add new JavaFX structure and FXML files

+++ capstone/src/main/java/de/fhdo/sama/capstone/model/Location.java
@@ -1,5 +1,3 @@
 package de.fhdo.sama.capstone.model;
 
-public record Location(String id, String name, int x, int y) {
-
-}
+public record Location(String id, String name, double x, double y) {}
